32bit compat

+++ pandas/tests/frame/indexing/test_where.py
@@ -724,7 +724,7 @@ class TestDataFrameIndexingWhere:
         # GH#45135, analogue to GH#44181 for Period don't raise on no-op
         # For td64/dt64/dt64tz we already don't raise, but also are
         #  checking that we don't unnecessarily upcast to object.
-        ser = Series(np.arange(3) * 10 ** 9).view(dtype)
+        ser = Series(np.arange(3) * 10 ** 9, dtype=np.int64).view(dtype)
         df = ser.to_frame()
         mask = np.array([False, False, False])
